+++ src/quickagram.js
@@ -1019,7 +1019,7 @@
 
   return {
     render,
-    version: '0.3.1',
+    version: '0.3.2',
     THEMES,
     SHAPES,
     autoLayout,